refactor: ModalContent 컴포넌트의 message prop명을 basicDescription로 변경
