+++ components/src/stories/AlertModal.stories.tsx
@@ -3,7 +3,7 @@ import { action } from '@storybook/addon-actions';
 
 import AlertModal from '../lib/components/Modal/AlertModal/index';
 
-import '../styles/index.css';
+import '../lib/styles/index.css';
 
 const meta: Meta<typeof AlertModal> = {
   title: 'Modal/Examples',
@@ -18,7 +18,7 @@ export default meta;
 export const Alert = AlertModal.bind({});
 Alert.args = {
   title: '확인(Alert) 모달',
-  message: '사용자에게 메시지를 전달하고 확인 버튼만 제공',
+  basicDescription: '사용자에게 메시지를 전달하고 확인 버튼만 제공',
   isCloseIcon: false,
   $modalSize: 'medium',
   $position: 'center',
